Introduce abstraction for component rendering functions and support for custom describe blocks
