+++ jest.conf/setup.js
@@ -10,6 +10,7 @@ import Vue from 'vue';
 import VueRouter from 'vue-router';
 import VueIntl from 'vue-intl';
 import KThemePlugin from '../lib/KThemePlugin';
+import { describeUnit, describeVisual } from './testUtils';
 
 global.beforeEach(() => {
   return new Promise(resolve => {
@@ -54,30 +55,5 @@ global.flushPromises = function flushPromises() {
   });
 };
 
-function removeWhitespaceFromHtml(htmlString) {
-  // https://stackoverflow.com/a/33108909
-  return htmlString.replace(/>\s+|\s+</g, function (m) {
-    return m.trim();
-  });
-}
-
-global.expect.extend({
-  // check that document.body.innerHTML includes html string
-
-  toBeInDom(received) {
-    const pass = removeWhitespaceFromHtml(document.body.innerHTML).includes(
-      removeWhitespaceFromHtml(received),
-    );
-    if (pass) {
-      return {
-        message: () => `expected ${received} not to be in the document body`,
-        pass: true,
-      };
-    } else {
-      return {
-        message: () => `expected ${received} to be in the document body`,
-        pass: false,
-      };
-    }
-  },
-});
+global.describeUnit = describeUnit;
+global.describeVisual = describeVisual;
